found a DRY way to set the og and twitter meta tags, very cool -> http://elegantbrew.tumblr.com/post/70990048275/controller-concerns-in-rails-4
