+++ app/controllers/about_controller.rb
@@ -1,4 +1,5 @@
 class AboutController < ApplicationController
+  include SetMetaTags
   skip_before_action :authenticate_user!
   before_action :set_meta_about
 
@@ -15,33 +16,10 @@ class AboutController < ApplicationController
     @page_title = 'about me'
     @page_description = 'About Pat Needham'
     @page_keywords = @page_keywords + keywords_about
-    set_meta_tags og: og_tags,
-                  twitter: twitter_tags
   end
 
   def keywords_about
     ['about', 'me', 'bio', 'professional', 'personal', 'opencity', 'ditto', 'general', 'assembly', 'wdi']
-  end
-
-  def og_tags
-    {
-      site_name:    'Pat Needham',
-      title:        'about me',
-      description:  'About Pat Needham',
-      image:        @image_src,
-      url:          request.env['REQUEST_URI']
-    }
-  end
-
-  def twitter_tags
-    {
-      card:         'summary',
-      site:         '@patneedham_',
-      title:        'about me',
-      description:  'About Pat Needham',
-      image:        @image_src,
-      url:          request.env['REQUEST_URI']
-    }
   end
 
 end
